+++ com.creative-scripts.cstk/jsx/1.jsx
@@ -1,1 +1,5 @@
-$.__fileName();
+alert(Math.random());
+// bug+2
+/*
+$.evalFile('C:\\Program Files\\Common Files\\Adobe\\CEP\\extensions\\CSTK\\jsx\1.jsx')
+ */